Robot climbs
Added a climbing function, if an object in the actor tag is labled Climbable the robot can climb it for about 1 second, using either the up or space button, this allows him to accomplish hights higher then being able to jump.  This is a function in the 2D character, run by the ontick.